+++ packages/plugins/app-plugin-ai-knowledge-base/client/providers/types.ts
@@ -19,7 +19,7 @@ export type KnowledgeBase = {
   vectorStoreProvider?: string;
   vectorStoreConfigKey?: string;
   vectorDatabaseKey?: string;
-  storageId?: string;
+  disk?: string;
   llmService?: string;
   embeddingModel?: string;
   vectorStoreProps?: Array<{ name?: string; key: string; value: unknown }>;
@@ -180,7 +180,7 @@ export type KnowledgeBaseMutation = {
   vectorStoreProvider?: string;
   vectorStoreConfigKey?: string;
   vectorDatabaseKey?: string;
-  storageId?: string;
+  disk?: string;
   llmService?: string;
   embeddingModel?: string;
   vectorStoreProps?: Array<{ name?: string; key: string; value: unknown }>;