+++ packages/templates/app-template-default/client/runtime.ts
@@ -2,7 +2,8 @@ import { createAppClientConfig } from '@nocobase/app-client';
 import { defineAppRuntime } from '@nocobase/app-client/runtime';
 import { getPortalBase } from '@nocobase/app-portal-sdk/runtime';
 
-import clientPlugins from './plugins.js';
+import locales from './locales/index.js';
+import plugins from './plugins.js';
 import reactProviders from './react-providers.js';
 import routeComponentOverrides from './route-overrides.js';
 import routes from './routes.js';
@@ -14,35 +15,12 @@ const appRuntime = defineAppRuntime({
   basename: getPortalBase(),
   config: createAppClientConfig,
   serviceProviders,
-  locales: {
-    'en-US': () => import('./locales/en-US.js'),
-    'zh-CN': () => import('./locales/zh-CN.js'),
-  },
+  locales,
   reactProviders,
   routes,
-  plugins: clientPlugins.plugins,
-  routeComponentOverrides: [
-    ...clientPlugins.routeComponentOverrides,
-    ...routeComponentOverrides,
-  ],
+  plugins,
+  routeComponentOverrides,
   sourceExtensions,
-  validate(app) {
-    if (!app.refineConfig.authProvider) {
-      throw new Error(
-        'Default App requires an enabled client plugin that registers an auth provider.',
-      );
-    }
-    if (
-      !app.runtime.routes.some(
-        (route) =>
-          route.path.toLowerCase() === '/login' && route.auth === 'guest',
-      )
-    ) {
-      throw new Error(
-        'Default App requires an enabled client plugin that registers a guest /login route.',
-      );
-    }
-  },
 });
 
 export default appRuntime;